add : 챗봇을 메인 화면에서 플로팅 액션 버튼을 눌렀을 때 뜨도록 수정함. 그리고 프래그먼트에 있었던 코드를 액티비티로 바꿈.
