+++ app/src/main/AndroidManifest.xml
@@ -26,9 +26,11 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
-            android:name=".ui.activity.login.LoginActivity"
+            android:name=".ui.activity.chatbot.ChatbotActivity"
             android:exported="false" />
-        <!-- <meta-data -->
+        <activity
+            android:name=".ui.activity.login.LoginActivity"
+            android:exported="false" /> <!-- <meta-data -->
         <!-- android:name="com.naver.maps.map.CLIENT_ID" -->
         <!-- android:value="@string/naver_map_sdk_client_id" /> -->
         <activity
