+++ app/src/main/java/com/ing/offroader/data/repository/SanListRepository.kt
@@ -138,7 +138,7 @@ class SanListRepository {
         if (documents != null) {
             for (i in documents.documents.indices) {
                 val sanInfo = documents.documents[i]
-                //////////
+                //
                 sanArrayList.add(
                     SanDTO(
                         sanImage = sanInfo["images"] as ArrayList<String>,
