+++ app/src/main/java/com/ing/offroader/ui/activity/chatbot/database/HikeyDao.kt
@@ -1,4 +1,4 @@
-package com.ing.offroader.ui.fragment.chatbot.database
+package com.ing.offroader.ui.activity.chatbot.database
 
 import androidx.lifecycle.LiveData
 import androidx.room.Dao
@@ -11,7 +11,7 @@ interface HikeyDao {
 
     // 하이키
     @Insert(onConflict = OnConflictStrategy.IGNORE) // INSERT, key 충돌이 나도 무시하고 뒤에 추가해주도록 함.
-    suspend fun insertHikeyChat(chat:HikeyData)
+    suspend fun insertHikeyChat(chat: HikeyData)
     @Query("SELECT * FROM hikey_data")
     fun getHikeyConversation(): LiveData<List<HikeyData>>
 
@@ -24,7 +24,7 @@ interface HikeyDao {
 interface BongbongDao {
     // 봉봉이
     @Insert(onConflict = OnConflictStrategy.IGNORE) // INSERT, key 충돌이 나도 무시하고 뒤에 추가해주도록 함.
-    suspend fun insertBongbongChat(chat:BongbongData)
+    suspend fun insertBongbongChat(chat: BongbongData)
     @Query("SELECT * FROM BONGBONG_DATA")
     fun getBongbongConversation(): LiveData<List<BongbongData>>
