+++ app/src/main/java/com/ing/offroader/ui/fragment/chatbot/MyApplication.kt
@@ -4,10 +4,10 @@ import android.app.Application
 import android.content.Context
 import com.ing.offroader.data.repository.EventRepository
 import com.ing.offroader.data.repository.SanListRepository
-import com.ing.offroader.ui.fragment.chatbot.database.BongbongDatabase
-import com.ing.offroader.ui.fragment.chatbot.database.HikeyDatabase
-import com.ing.offroader.ui.fragment.chatbot.viewmodel.BongbongRepository
-import com.ing.offroader.ui.fragment.chatbot.viewmodel.HikeyRepository
+import com.ing.offroader.ui.activity.chatbot.database.BongbongDatabase
+import com.ing.offroader.ui.activity.chatbot.database.HikeyDatabase
+import com.ing.offroader.data.repository.BongbongRepository
+import com.ing.offroader.data.repository.HikeyRepository
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.SupervisorJob
