+++ app/src/main/java/com/ing/offroader/ui/fragment/chatbot/viewmodel/ChatBotViewModelFactory.kt
@@ -3,6 +3,8 @@ package com.ing.offroader.ui.fragment.chatbot.viewmodel
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import com.ing.offroader.data.repository.AiRepository
+import com.ing.offroader.data.repository.BongbongRepository
+import com.ing.offroader.data.repository.HikeyRepository
 
 class ChatBotViewModelFactory(private val hikeyRepository: HikeyRepository, private val bongbongRepository: BongbongRepository): ViewModelProvider.Factory {
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
