+++ app/src/main/java/com/ing/offroader/ui/fragment/home/HomeFragment.kt
@@ -119,5 +119,4 @@ class HomeFragment : Fragment() {
         super.onDestroyView()
         _binding = null
     }
-
 }
