+++ app/src/main/res/layout/activity_main.xml
@@ -16,6 +16,8 @@
         app:layout_constraintBottom_toTopOf="@id/ll_bottom_play"
         app:layout_constraintTop_toTopOf="parent"/>
 
+
+
     <com.google.android.material.bottomnavigation.BottomNavigationView
         android:id="@+id/nav_main"
         android:layout_width="match_parent"
@@ -77,6 +79,15 @@
         app:layout_constraintTop_toTopOf="@id/tv_radio_title"
         android:layout_marginEnd="10dp"/>
 
+    <com.google.android.material.floatingactionbutton.FloatingActionButton
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        app:layout_constraintBottom_toTopOf="@id/nav_main"
+        android:layout_marginBottom="16dp"
+        app:layout_constraintEnd_toEndOf="parent"
+        android:layout_marginEnd="16dp"
+        android:id="@+id/fab_chatbot" />
+
     <LinearLayout
         android:id="@+id/ll_bottom_play"
         android:layout_width="match_parent"
@@ -84,6 +95,7 @@
         android:background="@color/white"
         android:orientation="horizontal"
         app:layout_constraintBottom_toTopOf="@id/nav_main">
+
 
         <androidx.cardview.widget.CardView
             android:layout_width="40dp"
